requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
attrs==25.4.0
boto3==1.40.11
botocore==1.40.11
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
cloudinary==1.44.1
cryptography==46.0.4
dj-database-url==3.0.1
Django==5.2.4
django-allauth==65.10.0
django-appconf==1.1.0
django-cleanup==9.0.0
django-cloudinary-storage==0.3.0
django-compressor==4.5.1
django-debug-toolbar==6.0.0
django-minio-storage==0.5.8
django-schema-viewer==0.5.3
django-storages==1.14.6
django-tailwind==4.2.0
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.10
jigsawstack==0.4.0
jmespath==1.0.1
minio==7.2.16
multidict==6.7.1
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pi==0.1.2
pillow==11.3.0
propcache==0.4.1
psycopg==3.2.9
psycopg2==2.9.10
pycparser==2.22
pycryptodome==3.23.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
rcssmin==1.1.2
requests==2.32.4
rjsmin==1.2.2
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
yarl==1.22.0
